Tried to fix conflicts from other PR, fixed some formatting on HTML
